Extract results cache into new pkg
