+++ pkg/logproto/compat.go
@@ -18,6 +18,7 @@ import (
 	"github.com/prometheus/prometheus/model/timestamp"
 
 	"github.com/grafana/loki/pkg/querier/queryrange/queryrangebase/definitions"
+	"github.com/grafana/loki/pkg/storage/chunk/cache/resultscache"
 	"github.com/grafana/loki/pkg/util"
 )
 
@@ -250,10 +251,10 @@ func (m *IndexStatsRequest) GetQuery() string {
 }
 
 // GetCachingOptions returns the caching options.
-func (m *IndexStatsRequest) GetCachingOptions() (res definitions.CachingOptions) { return }
+func (m *IndexStatsRequest) GetCachingOptions() (res resultscache.CachingOptions) { return }
 
 // WithStartEnd clone the current request with different start and end timestamp.
-func (m *IndexStatsRequest) WithStartEnd(start, end time.Time) definitions.Request {
+func (m *IndexStatsRequest) WithStartEnd(start, end time.Time) resultscache.Request {
 	clone := *m
 	clone.From = model.TimeFromUnixNano(start.UnixNano())
 	clone.Through = model.TimeFromUnixNano(end.UnixNano())
@@ -298,10 +299,10 @@ func (m *VolumeRequest) GetQuery() string {
 }
 
 // GetCachingOptions returns the caching options.
-func (m *VolumeRequest) GetCachingOptions() (res definitions.CachingOptions) { return }
+func (m *VolumeRequest) GetCachingOptions() (res resultscache.CachingOptions) { return }
 
 // WithStartEnd clone the current request with different start and end timestamp.
-func (m *VolumeRequest) WithStartEnd(start, end time.Time) definitions.Request {
+func (m *VolumeRequest) WithStartEnd(start, end time.Time) resultscache.Request {
 	clone := *m
 	clone.From = model.TimeFromUnixNano(start.UnixNano())
 	clone.Through = model.TimeFromUnixNano(end.UnixNano())
