+++ pkg/querier/queryrange/queryrangebase/alias.go
@@ -1,13 +1,19 @@
 package queryrangebase
 
-import "github.com/grafana/loki/pkg/querier/queryrange/queryrangebase/definitions"
+import (
+	"github.com/grafana/loki/pkg/querier/queryrange/queryrangebase/definitions"
+	"github.com/grafana/loki/pkg/storage/chunk/cache/resultscache"
+)
 
 // Helpful aliases for refactoring circular imports
 
-type CachingOptions = definitions.CachingOptions
+type CachingOptions = resultscache.CachingOptions
 type PrometheusResponseHeader = definitions.PrometheusResponseHeader
 type PrometheusRequestHeader = definitions.PrometheusRequestHeader
 type Codec = definitions.Codec
 type Merger = definitions.Merger
+type CacheGenNumberLoader = resultscache.CacheGenNumberLoader
+
 type Request = definitions.Request
 type Response = definitions.Response
+type Extent = resultscache.Extent
