+++ pkg/querier/queryrange/queryrangebase/definitions/interface.go
@@ -3,10 +3,10 @@ package definitions
 import (
 	"context"
 	"net/http"
-	"time"
 
-	"github.com/gogo/protobuf/proto"
 	"github.com/opentracing/opentracing-go"
+
+	"github.com/grafana/loki/pkg/storage/chunk/cache/resultscache"
 )
 
 // Codec is used to encode/decode query range requests and responses so they can be passed down to middlewares.
@@ -32,28 +32,16 @@ type Merger interface {
 
 // Request represents a query range request that can be process by middlewares.
 type Request interface {
-	// GetStart returns the start timestamp of the request in milliseconds.
-	GetStart() time.Time
-	// GetEnd returns the end timestamp of the request in milliseconds.
-	GetEnd() time.Time
-	// GetStep returns the step of the request in milliseconds.
-	GetStep() int64
-	// GetQuery returns the query of the request.
-	GetQuery() string
-	// GetCachingOptions returns the caching options.
-	GetCachingOptions() CachingOptions
-	// WithStartEnd clone the current request with different start and end timestamp.
-	WithStartEnd(start time.Time, end time.Time) Request
+	resultscache.Request
 	// WithQuery clone the current request with a different query.
 	WithQuery(string) Request
-	proto.Message
 	// LogToSpan writes information about this request to an OpenTracing span
 	LogToSpan(opentracing.Span)
 }
 
 // Response represents a query range response.
 type Response interface {
-	proto.Message
+	resultscache.Response
 	// GetHeaders returns the HTTP headers in the response.
 	GetHeaders() []*PrometheusResponseHeader
