+++ pkg/querier/queryrange/queryrangebase/queryrange.proto
@@ -3,11 +3,11 @@ syntax = "proto3";
 package queryrangebase;
 
 import "gogoproto/gogo.proto";
-import "google/protobuf/any.proto";
 import "google/protobuf/duration.proto";
 import "google/protobuf/timestamp.proto";
 import "pkg/logproto/logproto.proto";
 import "pkg/querier/queryrange/queryrangebase/definitions/definitions.proto";
+import "pkg/storage/chunk/cache/resultscache/types.proto";
 
 option go_package = "github.com/grafana/loki/pkg/querier/queryrange/queryrangebase";
 option (gogoproto.marshaler_all) = true;
@@ -29,7 +29,7 @@ message PrometheusRequest {
     (gogoproto.nullable) = false
   ];
   string query = 6;
-  definitions.CachingOptions cachingOptions = 7 [(gogoproto.nullable) = false];
+  resultscache.CachingOptions cachingOptions = 7 [(gogoproto.nullable) = false];
   repeated definitions.PrometheusRequestHeader Headers = 8 [(gogoproto.jsontag) = "-"];
 }
 
@@ -62,23 +62,4 @@ message SampleStream {
     (gogoproto.nullable) = false,
     (gogoproto.jsontag) = "values"
   ];
-}
-
-message CachedResponse {
-  string key = 1 [(gogoproto.jsontag) = "key"];
-
-  // List of cached responses; non-overlapping and in order.
-  repeated Extent extents = 2 [
-    (gogoproto.nullable) = false,
-    (gogoproto.jsontag) = "extents"
-  ];
-}
-
-message Extent {
-  int64 start = 1 [(gogoproto.jsontag) = "start"];
-  int64 end = 2 [(gogoproto.jsontag) = "end"];
-  // reserved the previous key to ensure cache transition
-  reserved 3;
-  string trace_id = 4 [(gogoproto.jsontag) = "-"];
-  google.protobuf.Any response = 5 [(gogoproto.jsontag) = "response"];
 }
