+++ pkg/querier/queryrange/queryrangebase/step_align.go
@@ -20,5 +20,5 @@ type stepAlign struct {
 func (s stepAlign) Do(ctx context.Context, r Request) (Response, error) {
 	start := (r.GetStart().UnixMilli() / r.GetStep()) * r.GetStep()
 	end := (r.GetEnd().UnixMilli() / r.GetStep()) * r.GetStep()
-	return s.next.Do(ctx, r.WithStartEnd(time.UnixMilli(start), time.UnixMilli(end)))
+	return s.next.Do(ctx, r.WithStartEnd(time.UnixMilli(start), time.UnixMilli(end)).(Request))
 }
