+++ pkg/querier/queryrange/roundtrip.go
@@ -7,11 +7,10 @@ import (
 	"strings"
 	"time"
 
-	"github.com/grafana/dskit/user"
-
 	"github.com/go-kit/log"
 	"github.com/go-kit/log/level"
 	"github.com/grafana/dskit/httpgrpc"
+	"github.com/grafana/dskit/user"
 	"github.com/pkg/errors"
 	"github.com/prometheus/client_golang/prometheus"
 	"github.com/prometheus/common/model"
