+++ pkg/querier/queryrange/volume_cache_test.go
@@ -10,6 +10,8 @@ import (
 	"github.com/prometheus/common/model"
 	"github.com/stretchr/testify/require"
 
+	"github.com/grafana/loki/pkg/storage/chunk/cache/resultscache"
+
 	"github.com/grafana/loki/pkg/logproto"
 	"github.com/grafana/loki/pkg/logqlmodel/stats"
 	"github.com/grafana/loki/pkg/querier/queryrange/queryrangebase"
@@ -22,8 +24,10 @@ import (
 func TestVolumeCache(t *testing.T) {
 	setup := func(volResp *VolumeResponse) (*int, queryrangebase.Handler) {
 		cfg := queryrangebase.ResultsCacheConfig{
-			CacheConfig: cache.Config{
-				Cache: cache.NewMockCache(),
+			Config: resultscache.Config{
+				CacheConfig: cache.Config{
+					Cache: cache.NewMockCache(),
+				},
 			},
 		}
 		c, err := cache.New(cfg.CacheConfig, nil, log.NewNopLogger(), stats.ResultCache, constants.Loki)
@@ -131,7 +135,7 @@ func TestVolumeCache(t *testing.T) {
 			},
 		}
 
-		resp, err = handler.Do(ctx, req)
+		resp, err = handler.Do(ctx, req.(queryrangebase.Request))
 		require.NoError(t, err)
 		require.Equal(t, 1, *calls)
 		require.Equal(t, expectedVol, resp)
@@ -281,8 +285,10 @@ func TestVolumeCache_RecentData(t *testing.T) {
 	} {
 		t.Run(tc.name, func(t *testing.T) {
 			cfg := queryrangebase.ResultsCacheConfig{
-				CacheConfig: cache.Config{
-					Cache: cache.NewMockCache(),
+				Config: resultscache.Config{
+					CacheConfig: cache.Config{
+						Cache: cache.NewMockCache(),
+					},
 				},
 			}
 			c, err := cache.New(cfg.CacheConfig, nil, log.NewNopLogger(), stats.ResultCache, constants.Loki)
