+++ pkg/storage/chunk/cache/resultscache/util.go
@@ -1,10 +1,15 @@
-package queryrangebase
+package resultscache
 
 import (
 	"context"
-
-	"github.com/grafana/loki/pkg/storage/chunk/cache/resultscache"
 )
+
+type HandlerFunc func(context.Context, Request) (Response, error)
+
+// Do implements Handler.
+func (q HandlerFunc) Do(ctx context.Context, req Request) (Response, error) {
+	return q(ctx, req)
+}
 
 // RequestResponse contains a request response and the respective request that was used.
 type RequestResponse struct {
@@ -59,24 +64,4 @@ func DoRequests(ctx context.Context, downstream Handler, reqs []Request, paralle
 	}
 
 	return resps, firstErr
-}
-
-type queryMergerAsCacheResponseMerger struct {
-	Merger
-}
-
-func (m *queryMergerAsCacheResponseMerger) MergeResponse(responses ...resultscache.Response) (resultscache.Response, error) {
-	cacheResponses := make([]Response, 0, len(responses))
-	for _, r := range responses {
-		cacheResponses = append(cacheResponses, r.(Response))
-	}
-	response, err := m.Merger.MergeResponse(cacheResponses...)
-	if err != nil {
-		return nil, err
-	}
-	return response.(resultscache.Response), nil
-}
-
-func FromMergerToCacheResponseMerger(m Merger) resultscache.ResponseMerger {
-	return &queryMergerAsCacheResponseMerger{m}
 }
